requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-crispy-forms==2.0
django-import-export==3.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
future==0.18.3
idna==3.4
MarkupPy==1.14
modeltranslation==0.25
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.0
pypng==0.20220715.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.4
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0